requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
attrs==25.3.0
bcrypt==5.0.0
beautifulsoup4==4.13.4
certifi==2024.6.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
cryptography==46.0.2
dataclasses-json==0.6.7
ddgs==9.5.2
distlib==0.3.8
distro==1.9.0
duckduckgo_search==8.1.1
ecdsa==0.19.1
fastapi==0.118.0
filelock==3.14.0
frozenlist==1.7.0
greenlet==3.2.3
griffe==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.7
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.10
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.98.0
openai-agents==0.0.3
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
passlib==1.7.4
platformdirs==4.2.2
primp==0.15.0
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.9.0
tools==1.0.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.2.1
uvicorn==0.37.0
virtualenv==20.26.2
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
